requirements.txt
Django
django-ckeditor
Pillow
django-widget-tweaks